feat(amenity): implement all method for amenity

+++ src/main/java/clyvasync/Clyvasync/dto/request/RoomAmenityHighlightRequest.java
@@ -4,6 +4,6 @@ import lombok.Data;
 
 @Data
 public class RoomAmenityHighlightRequest {
-    private Long amenityId;
+    private Integer amenityId;
     private String displayValue;
 }
