+++ src/main/java/clyvasync/Clyvasync/dto/request/UpdateHomestayAmenitiesRequest.java
@@ -6,5 +6,5 @@ import java.util.Set;
 
 @Data
 public class UpdateHomestayAmenitiesRequest {
-    private Set<Long> amenityIds;
+    private Set<Integer> amenityIds;
 }
